optimizer now works with single feature and one bias but has bugs with multi-feature data

+++ src/machine_learn/metrics/mean_squared_error.py
@@ -1,4 +1,4 @@
 from src.machine_learn.imports import np
 
 def mean_squared_error(y_pred, y_act):
-    return np.mean((y_pred-y_act)**2)
+    return float(np.mean((y_pred-y_act)**2))
